android-ibrdtn: Fix usage of getProcessingFlags() caused by recent C++ changes

+++ android/ibrdtn/src/de/tubs/ibr/dtn/service/ClientSession.java
@@ -652,7 +652,7 @@ public class ClientSession {
 		Block ret = new Block();
 		ret.type = Integer.valueOf(block.getType());
 		ret.length = block.getLength();
-		ret.procflags = block.getProcessingFlags();
+		ret.procflags = block.getProcessingFlags().longValue();
 		return ret;
 	}
